feat(modules): add endpoint to fetch Dapr port information
Add a new GET endpoint `/daprport` to the `auth` module in `index.ts` file. This endpoint fetches the Dapr port from the environment variable `DAPR_HTTP_PORT` and returns it in the response. If the port is successfully fetched, the response will have `success` set to `true`, `data` containing the `daprPort`, and a success message. If there is an error while fetching the port, the response will have `success` set to `false`, `data` set to `null`, and an error message.

fix(utils): remove console.log statement in dapr.ts

Remove the unnecessary `console.log` statement in the `saveToDaprState` function in `dapr.ts` file.