feat(db): add migration script for creating "User" table in the postgres-auth database
This commit adds a new migration script `migration.sql` in the `db/postgres-auth/migrations/20230910222904_academy_auth` directory. The script creates a table named "User" with the following columns:
- id: TEXT, NOT NULL
- firstname: VARCHAR(100), NOT NULL
- lastname: VARCHAR(100), NOT NULL
- email: VARCHAR(320), NOT NULL
- hash: VARCHAR(500), NOT NULL
- photo: VARCHAR(500)
- cellphone: VARCHAR(20)
- birthdate: TIMESTAMPTZ(6), NOT NULL
- createdAt: TIMESTAMP(3), NOT NULL, DEFAULT CURRENT_TIMESTAMP
- updatedAt: TIMESTAMP(3), NOT NULL
- status: TEXT, NOT NULL

The script also adds a primary key constraint on the "id" column and a unique index on the "email" column.

Additionally, this commit adds a new file `migration_lock.toml` in the `db/postgres-auth/migrations` directory. This file should not be manually edited and should be added to the version-control system (e.g., Git). The file specifies the provider as "postgresql".